api/requirements.txt
Django==2.2.4
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
PyYAML==5.1.2
uritemplate==3.0.0
psycopg2-binary>=2.7.3.2
django-redis>=4.10.0
gunicorn==19.9.0